[refactor] Move tests to test folder
